+++ src/Phan/LanguageServer/Protocol/MarkupContent.php
@@ -49,4 +49,11 @@ class MarkupContent
         $this->kind = $kind;
         $this->value = $value;
     }
+
+    public static function fromArray(array $data) : self {
+        return new self(
+            $data['kind'],
+            $data['value']
+        );
+    }
 }